Rename command git-update to update-rsync and enable functionality
